refactor: Rename rst module to rst_document
