+++ varst/utils/rst_document.py
